+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.9.9 / 2014-05-xx
 ==================
 
+* Bugfix: properly respect resolutionStrategy.force 'group:artifact:version' present in build.gradle
 * Add commitLock task
 
 1.9.8 / 2014-05-02